lots of changes fixed all the position index look ups - it actually seems to somewhat work moved to float 64 array - however it isn't accurate enough I don't think - so attempting to convert to position decoding

+++ Source/Core/HeightmapTessellator.js
@@ -614,7 +614,7 @@ HeightmapTessellator.computeVertices = function (options) {
   var transform;
   var inverseTransform;
 
-  var packedPositions = new Float32Array(positions.length * 3);
+  var packedPositions = new Float64Array(positions.length * 3);
   for (var i = 0; i < positions.length; i++) {
     var pos = positions[i];
     packedPositions[i * 3] = pos.x;
